REFACTOR :: file path, instance host, user authority
