+++ src/allFiles.tsx
@@ -3,12 +3,12 @@ import Header from "./components/Header";
 import Main from "./components/Main";
 import Teacher from "./components/Teacher";
 import TeacherTier from "./components/teacherDetails/TeacherTier";
-import Comment from "./components/comment";
-import Recomment from "./components/Recomment";
+import Comment from "./components/comment/comment";
+import Recomment from "./components/comment/Recomment";
 import Survey from "./components/Survey";
 import SurveyList from "./components/SurveyList";
-import Login from "./components/Login";
-import Signup from "./components/Signup";
+import Login from "./components/auth/Login";
+import Signup from "./components/auth/Signup";
 
 export {
   TeacherInfo,
