+++ src/instance.js
@@ -1,5 +1,5 @@
 import axios from "axios";
 
 export const instance = axios.create({
-  baseURL: "http://10.150.151.125:8080",
+  baseURL: "http://15.164.246.49:8080",
 });
